Clean up main, Use byte type instead of int for pins. Closes #4 and closes #8

+++ include/constants.h
@@ -1,33 +1,33 @@
 #pragma once
 namespace pin {
     // display
-    static const int disp_digit1 = 22;
-    static const int disp_digit2 = 23;
-    static const int disp_digit3 = 24;
-    static const int disp_digit4 = 25;
+    static const byte disp_digit1 = 22;
+    static const byte disp_digit2 = 23;
+    static const byte disp_digit3 = 24;
+    static const byte disp_digit4 = 25;
 
-    static const int disp_segA = 26;
-    static const int disp_segB = 27;
-    static const int disp_segC = 28;
-    static const int disp_segD = 29;
-    static const int disp_segE = 30;
-    static const int disp_segF = 31;
-    static const int disp_segG = 32;
-    static const int disp_segP = 33;
+    static const byte disp_segA = 26;
+    static const byte disp_segB = 27;
+    static const byte disp_segC = 28;
+    static const byte disp_segD = 29;
+    static const byte disp_segE = 30;
+    static const byte disp_segF = 31;
+    static const byte disp_segG = 32;
+    static const byte disp_segP = 33;
 
     // input
-    static const int in_button1 = 52;
-    static const int in_button2 = 53;
+    static const byte in_button1 = 52;
+    static const byte in_button2 = 53;
 
     // LEDs
-    static const int led_statusR = 2;
-    static const int led_statusG = 3;
-    static const int led_statusB = 4;
-    static const int led_countA = 48;
-    static const int led_countB = 49;
-    static const int led_countC = 50;
-    static const int led_countD = 51;
+    static const byte led_statusR = 2;
+    static const byte led_statusG = 3;
+    static const byte led_statusB = 4;
+    static const byte led_countA = 48;
+    static const byte led_countB = 49;
+    static const byte led_countC = 50;
+    static const byte led_countD = 51;
 
     // audio
-    static const int audio_ticker = 5;
+    static const byte audio_ticker = 5;
 }
